+++ hutool-extra/src/main/java/cn/hutool/extra/ssh/Sftp.java
@@ -238,7 +238,10 @@ public class Sftp extends AbstractFtp {
 				return LsEntrySelector.CONTINUE;
 			});
 		} catch (SftpException e) {
-			throw new JschRuntimeException(e);
+			if(false == StrUtil.startWithIgnoreCase(e.getMessage(), "No such file")){
+				throw new JschRuntimeException(e);
+			}
+			// 文件不存在忽略
 		}
 		return fileNames;
 	}